+++ lib/kickstart/accounts.ex
@@ -443,6 +443,82 @@ defmodule Kickstart.Accounts do
     end
   end
 
+  # Social login
+
+  def find_or_create(%{} = auth) do
+    user =
+      email_from_auth(auth)
+      |> get_user_for_email()
+
+    if user do
+      {:ok, user}
+    else
+      create_user(basic_info(auth))
+    end
+  end
+
+  defp get_user_for_email(email) do
+    IO.inspect(Repo.get_by(User, email: email))
+  end
+
+  # github does it this way
+  defp avatar_from_auth(%{info: %{urls: %{avatar_url: image}}}), do: image
+
+  # facebook does it this way
+  defp avatar_from_auth(%{info: %{image: image}}), do: image
+
+  # default case if nothing matches
+  defp avatar_from_auth(auth) do
+    require Logger
+    Logger.warn("#{auth.provider} needs to find an avatar URL!")
+    Logger.debug(Jason.encode!(auth))
+    nil
+  end
+
+  defp email_from_auth(%{info: %{email: email}}), do: email
+
+  defp basic_info(auth) do
+    email = email_from_auth(auth)
+    random_password = Ecto.UUID.generate |> binary_part(16,12)
+
+    case auth.strategy do
+      Ueberauth.Strategy.Facebook ->
+        %{
+          social_auth_uid: auth.uid,
+          name: name_from_auth(auth),
+          email: email,
+          avatar: avatar_from_auth(auth),
+          provider: "facebook",
+          password: random_password
+        }
+
+      Ueberauth.Strategy.Google ->
+        %{
+          social_auth_uid: auth.uid,
+          name: name_from_auth(auth),
+          email: email,
+          avatar: avatar_from_auth(auth),
+          provider: "google",
+          password: random_password
+        }
+    end
+  end
+
+  defp name_from_auth(auth) do
+    if auth.info.name do
+      auth.info.name
+    else
+      name =
+        [auth.info.first_name, auth.info.last_name]
+        |> Enum.filter(&(&1 != nil and &1 != ""))
+
+      if Enum.empty?(name) do
+        auth.info.nickname
+      else
+        Enum.join(name, " ")
+      end
+    end
+  end
 
   # Torch code below
   import Torch.Helpers, only: [sort: 1, paginate: 4]