+++ priv/repo/migrations/20201217155638_add_social_auth_to_users.exs
@@ -1,0 +1,11 @@
+defmodule Kickstart.Repo.Migrations.AddSocialAuthToUsers do
+  use Ecto.Migration
+
+  def change do
+    alter table(:users) do
+      add :provider, :string
+      add :social_auth_uid, :string
+      add :avatar, :string
+    end
+  end
+end